requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0

# Google Cloud
google-cloud-storage==2.14.0
google-cloud-firestore==2.14.0
google-cloud-documentai==2.24.0
google-cloud-secret-manager==2.18.0
firebase-admin==6.4.0

# PDF Processing
PyMuPDF==1.23.22
Pillow==10.2.0
python-magic==0.4.27

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
httpx==0.26.0
tenacity==8.2.3
structlog==24.1.0

# Development
pytest>=7.0.0,<8.0.0
pytest-asyncio>=0.21.0
black==24.1.1
ruff==0.2.1
